refactor: rename deterministic-insights to activity-summarizer for clarity
Replace ambiguous implementation-focused filename with domain-focused descriptor to improve codebase readability.

TAG=agy
CONV=bd5d3037-c61b-4480-b219-1fa0d96ef1db
